[HUDI-4471] Relocate AWSDmsAvroPayload class to hudi-common
