+++ hudi-common/src/main/java/org/apache/hudi/common/model/AWSDmsAvroPayload.java
@@ -16,9 +16,8 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.payload;
+package org.apache.hudi.common.model;
 
-import org.apache.hudi.common.model.OverwriteWithLatestAvroPayload;
 import org.apache.hudi.common.util.Option;
 
 import org.apache.avro.Schema;
